+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This file is used to list changes made in each version of the haproxy cookbook.
 
 ## Unreleased
 
+- Add provides to service resource
+
 ## [v7.0.0] (2019-04-10)
 
 - Add `health` to allowed values for `mode` on `frontend`, `backend`, `listen`, `default`